Created exceptions package and refactored updateStatus to use streams
