+++ src/main/java/lithium/university/exceptions/TwitterServiceException.java
@@ -1,4 +1,4 @@
-package lithium.university.services;
+package lithium.university.exceptions;
 
 public class TwitterServiceException extends Exception {
     public TwitterServiceException(String message) {
