+++ src/main/java/lithium/university/services/TwitterService.java
@@ -1,6 +1,7 @@
 package lithium.university.services;
 
 import lithium.university.TwitterProperties;
+import lithium.university.exceptions.TwitterServiceException;
 import lithium.university.models.TwitterPost;
 import lithium.university.models.TwitterUser;
 import org.slf4j.Logger;
@@ -14,6 +15,7 @@ import twitter4j.conf.ConfigurationBuilder;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class TwitterService {
     private static final TwitterService INSTANCE = new TwitterService();
@@ -38,9 +40,8 @@ public class TwitterService {
             throw new TwitterServiceException("Cannot post. Message length should not exceed 280 characters");
         }
 
-        Status status = twitter.updateStatus(message);
-        logger.info("Successfully updated status to: \"" + status.getText() + "\"");
-        return status;
+        logger.info("Successfully updated status");
+        return Stream.of(twitter.updateStatus(message)).collect(Collectors.toList()).get(0);
     }
 
     /*
@@ -57,8 +58,11 @@ public class TwitterService {
     }
 
     public List<TwitterPost> retrieveFilteredFromTwitter(Twitter twitter, final int tweetTotal, String keyword) throws TwitterException {
+        if(keyword == null) {
+            return null;
+        }
         Paging p = new Paging(1, tweetTotal);
-        logger.debug("Attempting to grab " + tweetTotal + " tweets from Twitter timeline");
+        logger.debug("Attempting find tweets from Twitter timeline that match keyword: " + keyword);
         List<Status> statuses = twitter.getHomeTimeline(p);
         return statuses.stream().filter(s -> s.getText().contains(keyword))
                                 .map(s -> new TwitterPost(s.getText(),
