+++ src/test/java/lithium/university/TwitterServiceTest.java
@@ -1,8 +1,8 @@
 package lithium.university;
 
+import lithium.university.exceptions.TwitterServiceException;
 import lithium.university.models.TwitterPost;
 import lithium.university.services.TwitterService;
-import lithium.university.services.TwitterServiceException;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.Assert;
 import org.junit.Before;
@@ -156,6 +156,11 @@ public class TwitterServiceTest {
         List<TwitterPost> l = twitterServiceTest.retrieveFilteredFromTwitter(twitterTest, 1, "1");
         Assert.assertEquals(1, l.size());
         Assert.assertEquals("Tester 1", l.get(0).getTwitterMessage());
+    }
+
+    @Test
+    public void testRetrieveFilterNullKeyword() throws TwitterException {
+        Assert.assertNull(twitterServiceTest.retrieveFilteredFromTwitter(twitterTest, 1, null));
     }
 
     @Test
